demo/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
future==0.18.2
idna==2.8
oauth2-provider==0.0
oauthlib==3.1.0
pkg-resources==0.0.0
PySocks==1.7.1
python-twitter==3.5
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
tweepy==3.8.0
urllib3==1.25.7
